Load reading links from sqlite3 db
This is copied straight from our ArchiveBox install for now and will be
hosted in DO Spaces. The scripts downloads it from there, reads the
data it needs, then cleans up the file. We'll just need to upload new
dbs to spaces to keep everything updates.

Also fix some issues in the resume due to bumping `date-fns`.